feat: criando estrutura do projeto
