+++ .gitignore
@@ -1,0 +1,5 @@
+# IntelliJ related
+*.iml
+*.ipr
+*.iws
+.idea/
