+++ guia_unb/lib/core/config/theme/app_theme.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class AppTheme {
+  static ThemeData appTheme() {
+    return ThemeData(
+      fontFamily: 'Poppins',
+      textTheme: const TextTheme(
+        labelMedium: TextStyle(
+        ),
+        titleLarge: TextStyle(
+          fontWeight: FontWeight.bold,
+        ),
+      ),
+    );
+  }
+}
