+++ guia_unb/lib/pages/onboarding/onboarding.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class Onboarding extends StatelessWidget {
+  const Onboarding({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Stack(
+        children: [
+          Image.asset(
+            "assets/images/logo.png",
+            fit: BoxFit.cover,
+            width: double.infinity,
+            height: double.infinity,
+          ),
+        ],
+      ),
+    );
+  }
+}
